backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
